Move the block definitions to [blockname].block.json files
