+++ assets/course-theme/blocks/course-navigation/course-navigation.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/course-navigation",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
