+++ assets/course-theme/blocks/lesson-blocks/course-content.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/course-content",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
