+++ assets/course-theme/blocks/lesson-blocks/course-theme-course-progress-bar.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/course-theme-course-progress-bar",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
