+++ assets/course-theme/blocks/lesson-blocks/course-theme-course-progress-counter.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/course-theme-course-progress-counter",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
