+++ assets/course-theme/blocks/lesson-blocks/course-theme-lesson-actions.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/course-theme-lesson-actions",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
