+++ assets/course-theme/blocks/lesson-blocks/course-theme-lesson-module.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/course-theme-lesson-module",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
