+++ assets/course-theme/blocks/lesson-blocks/course-theme-notices.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/notices",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
