+++ assets/course-theme/blocks/lesson-blocks/course-theme-post-title.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/course-theme-post-title",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
