+++ assets/course-theme/blocks/lesson-blocks/course-theme-prev-next-lesson.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/course-theme-prev-next-lesson",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
