+++ assets/course-theme/blocks/lesson-blocks/course-title.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/course-title",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
