+++ assets/course-theme/blocks/lesson-blocks/exit-course.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/exit-course",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
