+++ assets/course-theme/blocks/lesson-blocks/focus-mode-toggle.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/focus-mode-toggle",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
