+++ assets/course-theme/blocks/lesson-blocks/page-actions.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/page-actions",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
