+++ assets/course-theme/blocks/lesson-blocks/sidebar-toggle-button.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/sidebar-toggle-button",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
