+++ assets/course-theme/blocks/quiz-blocks/quiz-actions.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/quiz-actions",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
