+++ assets/course-theme/blocks/quiz-blocks/quiz-back-to-lesson.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/quiz-back-to-lesson",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
