+++ assets/course-theme/blocks/quiz-blocks/quiz-progress.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/quiz-progress",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
