+++ assets/course-theme/blocks/ui/spacer.block.json
@@ -1,0 +1,6 @@
+{
+  "name": "sensei-lms/spacer-flex",
+  "category": "theme",
+  "apiVersion": 2,
+  "supports": {}
+}
